# Slightly improve performance of zend_copy_extra_args