@dstogov 2025-03-26
I'm not sure if the improvement comes from a single eliminated load from L1.
It may be the effect of the different code alignment. The presentation, that I sent you yesterday, talk about up to 30% difference on short loops caused by bad/good alignment.

In my tests I see ~2% improvement on your test above, and slight degradation on bench.php (both real-time and callgrind). This  is because ``i_init_func_execute_data()`` was optimized for case without extra args (that is used more often).

I'm indifferent to this.