@nielsdos 2025-04-02
Given that, and that there are ideas to implement some functions in PHP itself (which would also avoid this overhead when JITted for example), I'll leave this for now.